Refactor object tracker classes to remove unused utils.hpp
