+++ perception/autoware_multi_object_tracker/lib/association/association.cpp
@@ -17,7 +17,6 @@
 #include "autoware/multi_object_tracker/association/solver/gnn_solver.hpp"
 #include "autoware/multi_object_tracker/object_model/shapes.hpp"
 #include "autoware/multi_object_tracker/object_model/types.hpp"
-#include "autoware/multi_object_tracker/utils/utils.hpp"
 
 #include <autoware/object_recognition_utils/object_recognition_utils.hpp>
