+++ perception/autoware_multi_object_tracker/lib/tracker/model/pass_through_tracker.cpp
@@ -18,8 +18,6 @@
 #define EIGEN_MPL2_ONLY
 #include "autoware/multi_object_tracker/tracker/model/pass_through_tracker.hpp"
 
-#include "autoware/multi_object_tracker/utils/utils.hpp"
-
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 #include <autoware/object_recognition_utils/object_recognition_utils.hpp>
