+++ perception/autoware_multi_object_tracker/lib/tracker/model/pedestrian_tracker.cpp
@@ -19,8 +19,6 @@
 
 #include "autoware/multi_object_tracker/tracker/model/pedestrian_tracker.hpp"
 
-#include "autoware/multi_object_tracker/utils/utils.hpp"
-
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 #include <autoware/object_recognition_utils/object_recognition_utils.hpp>
