+++ perception/autoware_multi_object_tracker/lib/tracker/model/tracker_base.cpp
@@ -16,8 +16,6 @@
 
 #include "autoware/multi_object_tracker/tracker/model/tracker_base.hpp"
 
-#include "autoware/multi_object_tracker/utils/utils.hpp"
-
 #include <algorithm>
 #include <random>
 #include <vector>
