+++ perception/autoware_multi_object_tracker/src/multi_object_tracker_node.cpp
@@ -19,7 +19,6 @@
 
 #include "autoware/multi_object_tracker/object_model/shapes.hpp"
 #include "autoware/multi_object_tracker/uncertainty/uncertainty_processor.hpp"
-#include "autoware/multi_object_tracker/utils/utils.hpp"
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
